+++ src/main/java/org/ahpuh/surf/post/controller/PostController.java
@@ -4,14 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.ahpuh.surf.category.dto.CategorySimpleDto;
 import org.ahpuh.surf.common.cursor.CursorResult;
 import org.ahpuh.surf.jwt.JwtAuthentication;
-import org.ahpuh.surf.post.dto.ExploreDto;
-import org.ahpuh.surf.post.dto.PostCountDto;
-import org.ahpuh.surf.post.dto.RecentPostDto;
 import org.ahpuh.surf.post.dto.request.PostRequestDto;
-import org.ahpuh.surf.post.dto.response.AllPostResponseDto;
-import org.ahpuh.surf.post.dto.response.PostReadResponseDto;
-import org.ahpuh.surf.post.dto.response.PostResponseDto;
-import org.ahpuh.surf.post.dto.response.PostsRecentScoreResponseDto;
+import org.ahpuh.surf.post.dto.response.*;
 import org.ahpuh.surf.post.service.PostService;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
