+++ src/main/java/org/ahpuh/surf/post/domain/repository/PostRepositoryQuerydsl.java
@@ -1,9 +1,10 @@
 package org.ahpuh.surf.post.domain.repository;
 
-import org.ahpuh.surf.post.dto.ExploreDto;
-import org.ahpuh.surf.post.dto.PostCountDto;
 import org.ahpuh.surf.post.dto.PostScoreCategoryDto;
 import org.ahpuh.surf.post.dto.response.AllPostResponseDto;
+import org.ahpuh.surf.post.dto.response.ExploreResponseDto;
+import org.ahpuh.surf.post.dto.response.PostCountResponseDto;
+import org.ahpuh.surf.post.dto.response.RecentPostResponseDto;
 import org.ahpuh.surf.user.domain.User;
 import org.springframework.data.domain.Pageable;
