+++ src/main/java/org/ahpuh/surf/post/dto/response/ExploreResponseDto.java
@@ -1,4 +1,4 @@
-package org.ahpuh.surf.post.dto;
+package org.ahpuh.surf.post.dto.response;
 
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.*;
@@ -10,54 +10,37 @@ import java.time.LocalDateTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class RecentPostDto {
+public class ExploreResponseDto {
 
     private Long userId;
-
     private String userName;
-
     private String profilePhotoUrl;
-
-    @Builder.Default
-    private boolean isFollowedUser = false;
-
     private String categoryName;
-
     private String colorCode;
-
     private Long postId;
-
     private String content;
-
     private Integer score;
-
     private String imageUrl;
-
     private String fileUrl;
-
     private LocalDate selectedDate;
-
     private LocalDateTime createdAt;
-
-    @Builder.Default
-    private Long likeId = null;
-
-    @Builder.Default
-    private boolean isLiked = false;
+    private Long likeId;
+    private Boolean isLiked;
 
     @QueryProjection
-    public RecentPostDto(final Long userId,
-                         final String userName,
-                         final String profilePhotoUrl,
-                         final String categoryName,
-                         final String colorCode,
-                         final Long postId,
-                         final String content,
-                         final Integer score,
-                         final String imageUrl,
-                         final String fileUrl,
-                         final LocalDate selectedDate,
-                         final LocalDateTime createdAt) {
+    public ExploreResponseDto(final Long userId,
+                              final String userName,
+                              final String profilePhotoUrl,
+                              final String categoryName,
+                              final String colorCode,
+                              final Long postId,
+                              final String content,
+                              final Integer score,
+                              final String imageUrl,
+                              final String fileUrl,
+                              final LocalDate selectedDate,
+                              final LocalDateTime createdAt,
+                              final Long likeId) {
         this.userId = userId;
         this.userName = userName;
         this.profilePhotoUrl = profilePhotoUrl;
@@ -70,16 +53,11 @@ public class RecentPostDto {
         this.fileUrl = fileUrl;
         this.selectedDate = selectedDate;
         this.createdAt = createdAt;
-        this.likeId = null;
+        this.likeId = likeId;
         this.isLiked = false;
     }
 
-    public void setLiked(final Long likeId) {
-        this.likeId = likeId;
-        this.isLiked = true;
-    }
-
-    public void checkFollowed() {
-        this.isFollowedUser = true;
+    public void likeCheck() {
+        isLiked = likeId != null;
     }
 }
