+++ src/main/java/org/ahpuh/surf/post/service/PostService.java
@@ -17,15 +17,9 @@ import org.ahpuh.surf.common.exception.user.UserNotFoundException;
 import org.ahpuh.surf.post.domain.Post;
 import org.ahpuh.surf.post.domain.PostConverter;
 import org.ahpuh.surf.post.domain.repository.PostRepository;
-import org.ahpuh.surf.post.dto.ExploreDto;
-import org.ahpuh.surf.post.dto.PostCountDto;
 import org.ahpuh.surf.post.dto.PostScoreCategoryDto;
-import org.ahpuh.surf.post.dto.RecentPostDto;
 import org.ahpuh.surf.post.dto.request.PostRequestDto;
-import org.ahpuh.surf.post.dto.response.AllPostResponseDto;
-import org.ahpuh.surf.post.dto.response.PostReadResponseDto;
-import org.ahpuh.surf.post.dto.response.PostResponseDto;
-import org.ahpuh.surf.post.dto.response.PostsRecentScoreResponseDto;
+import org.ahpuh.surf.post.dto.response.*;
 import org.ahpuh.surf.s3.FileStatus;
 import org.ahpuh.surf.s3.S3Service;
 import org.ahpuh.surf.user.domain.User;
@@ -38,7 +32,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Slf4j
